Replace schedule_rebroadcast.py with a script for guided manual testing
